Purge.js
This command does exactly like the name says, purges messages by users!
Extra features: 
Doesn't give an annoying message of how many messages it has deleted.
If a user with the right permissions tries to use the command it deletes the command by the user. Leaving a clean chat behind!